Really disable SELinux
Historically /etc/sysconfig/selinux was a symlink to /etc/selinux/config, so you could modify one of the two files and you'd be done with it.

For an unknow reason this is no longer the case on CentOS 7. As they have exactly the same content by default it's not obvious that there was any change.

/etc/sysconfig/selinux seems to be ignored now, as changes have no action on the state of SELinux after a reboot.