Fix issues with env variable values containing special characters
When storing variables, their values are not sanitized. If they contain the same character as the one used for sed's field separator, everything goes wrong.

(and hello all the underhanded manipulations that could be done with that!)

Instead of replacing the value, delete the line and append it to the file again.