Make the host repo GPG keys available for image creation
By default, ``rpm --import`` adds the keys to the local rpm db (check with ``rpm -qi gpg-pubkey``) but doesn't copy them to ``/etc/pki/rpm-gpg/``.

We need those keys for the image creation, so copy them to the directory too, and bind mount it.

And do a ``yum update`` earlier because the non-updated rpm is somewhat buggy.