Temporary hack to work around slurm-pre uid/gid conflicts on images
Currently all postscript packages are installed in an image before
postscripts are actually run. Consequently RPMs setup user/group IDs
before postscripts that are designed for that specific purpose get
to do their thing.
This a temporary workaround to solve the issue for slurm-pre. It will
be removed after the upcoming image-creation process overhaul.