Add OpenSSH packages for the node images
In the case where we don't have a group file definition when creating the local repo and then we do an install from a local repository, some packages will be missing. The most important of them is openssh-server.

This is a workaround for that issue until we figure out a way to deal with package groups.