Yum and rpm management functions
Those functions work around some issues with the return codes of both yum and rpm.

When using them, you will get a helpful error message if some packages were not able to be installed by yum, as well as the messages from rpm if there was an issue at some point.