Rewrite parts of the galera resource agent to handle a two-node Master/Slave setup
- When the galera cluster is booting up mariadb will not start unless
both nodes are available and can report their respective seqno (indicating
the last commit). The RA will then be able to decide which node to promote
to the master role.

- In case only one node is available and it is desired to start mariadb, an admin
can force the bootstrap by running this command on that node:

crm_attribute -l reboot --name "<resource_name>-bootstrap" -v "true"

When the second node is later booted up, it will simply join the cluster
as a slave node.