NFS-server: don't use the separate rpcpipefs_dir
TL;DR: would cause major fsck-ups.

Using a separate rpcpipefs_dir causes problems when stopping the NFS server. One of the bind umounts done by the resource scripts fails, either due to timing issues, or to the resource script umounting in the same order as it mounted (bad idea!), or no second umount even happening, go figure. The net result is that the shared device is still mounted because of that leftover bind mount, even if the trinity-fs resource stopped.

It's not really a problem when the cause of the failover is the crash of the primary, or when it's properly fenced and taken out. For just a role change between two live systems though, this would be one heck of an issue.

The potential benefits of a separate rpcpipefs directory on a local disk are thin at best (maybe a bit faster?), so it's just not worth the trouble.