Shared-storage: remove problematic resource cleanup
I added that before I had the functions to check the state of resources, and there were timing issues that forced me to cleanup resources to restart them.

Now that cleanup and the following restart of the resources got my test clusters stuck a couple of times. It seems that they're waiting for a DRBD resource to get promoted, but as it's already been promoted in the past, nothing is happening and the resources don't come up.

The simplest solution for that is: pcs cluster stop && pcs cluster start

But that's not very pretty.