Update slurm-user.conf to have a more sensible QoS weight
PriorityWeightQOS in SLURM (like all the weights) is multiplied by a float between 0 and 1 and then cast back into an unsigned integer. The old default of one will truncate back to zero in all case. Updated this to default to 10000 and added a comment that all should be similar numbers.