New release of TrinityX^2.
A overhaul of the code has been done and parts of the code have been
removed to make way for a redesign on high-availability and security.

This release introduces the support for Enterprise Linux 8 and 9
and Ubuntu 22. Enterprise Linux 7 is deprecated.

The major change is Luna2 which an API driven and designed from scratch.
All supporting scripts and components have been rewritten in python3.